feat: add pkg/backoff sub-module for shared backoff strategies
Extract backoff implementations into a new importable sub-module at
pkg/backoff/ so external consumers (e.g. Horde) can use the same
algorithms instead of porting copies. Adds a Duration() method that
returns the next wait and advances state without blocking, for callers
that manage their own timers. The internal package now re-exports from
the public module.
