+++ pkg/backoff/backoff_test.go
@@ -138,3 +138,56 @@ func TestReset(t *testing.T) {
 		})
 	}
 }
+
+func TestDuration(t *testing.T) {
+	init := time.Millisecond
+	max := 5 * time.Second
+
+	t.Run("EqualJitterBackoff", func(t *testing.T) {
+		bo := &EqualJitterBackoff{
+			init:   init,
+			max:    max,
+			randFn: func(t time.Duration) time.Duration { return t },
+		}
+		bo.Reset()
+
+		d1 := bo.Duration()
+		d2 := bo.Duration()
+		assert.Less(t, d1, d2, "duration should increase")
+
+		bo.Reset()
+		d3 := bo.Duration()
+		assert.Equal(t, d1, d3, "duration should reset to initial value")
+	})
+
+	t.Run("ExpBackoff", func(t *testing.T) {
+		bo := &ExpBackoff{
+			duration: init,
+			init:     init,
+			max:      max,
+		}
+
+		d1 := bo.Duration()
+		d2 := bo.Duration()
+		assert.Less(t, d1, d2, "duration should increase")
+
+		bo.Reset()
+		d3 := bo.Duration()
+		assert.Equal(t, d1, d3, "duration should reset to initial value")
+	})
+}
+
+func TestNewEqualJitterBackoffWithDone(t *testing.T) {
+	init := time.Millisecond
+	max := 5 * time.Second
+
+	bo := NewEqualJitterBackoffWithDone(init, max)
+	assert.NotNil(t, bo)
+
+	d := bo.Duration()
+	assert.Greater(t, d, time.Duration(0), "duration should be positive")
+
+	bo.Reset()
+	d2 := bo.Duration()
+	assert.Greater(t, d2, time.Duration(0), "duration after reset should be positive")
+}
