+++ pkg/backoff/go.mod
@@ -1,0 +1,11 @@
+module github.com/elastic/elastic-agent/pkg/backoff
+
+go 1.25.9
+
+require github.com/stretchr/testify v1.9.0
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
